+++ CHANGELOG.md
@@ -6,6 +6,8 @@ These are the changes that have not been released yet. The release notes have no
 
 **FEATURE ENCHANCEMENTS:**
 
+- Re-add configuration file `podmod.conf`.
+
 - Add `--no-prune` option to `build`.
 
 - Add `--idempotent` option to `build`, `load`, and `unload`.